requirements.txt
-e .
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
requests==2.29.0
numpy==1.24.3
python-dotenv==1.0.0
pymongo==4.3.3
datasets==2.12.0
python-multipart==0.0.6
